Attendance Management system
An Attendance Management System (AMS) is an application designed to track and manage the attendance of individuals, typically in an educational or organizational setting. Functionality to mark attendance manually and Storing attendance data in a database . It is login for students, teachers, and administrators. I will handle the generate attendance reports for individuals or groups.